refactor: move analytics page sections to sections component
